+++ demo-app/src/example/Example.react.js
@@ -133,9 +133,9 @@ export default class Example extends React.Component<Props, State> {
             <fbt desc="header">Construct sentences</fbt>
           </h1>
           <h2>
-            <fbt desc="yet another header">
-              Use the form below to see FBT in action.
-            </fbt>
+            {/* For fbt common strings, the description will be sourced from an external manifest.
+            See `--fbt-common-path` option from `fbt-collect` and common_strings.json */}
+            <fbt common={true}>Use the form below to see FBT in action.</fbt>
           </h2>
           <form action="" method="get" onSubmit={this.onSubmit}>
             <fieldset>