+++ demo-app/webpack.config.js
@@ -54,9 +54,10 @@ module.exports = {
               '@babel/plugin-proposal-optional-chaining',
               '@babel/plugin-transform-flow-strip-types',
               ['babel-plugin-fbt', {
+                fbtCommonPath: './common_strings.json',
                 // We can also provide the fbt enum manifest directly as a JS variable
                 // fbtEnumManifest: require('./.enum_manifest.json'),
-                fbtEnumPath: path.join(__dirname, '.enum_manifest.json')
+                fbtEnumPath: path.join(__dirname, '.enum_manifest.json'),
               }],
               'babel-plugin-fbt-runtime',
             ]